modifications to f2c.h to conform to the LP64 data model in which int should be 32bit, not 64.  See the typedef for integer and ftnint and related in f2c.h

+++ CLAPACK-3.2.1/F2CLIBS/libf2c/f2c.h
@@ -7,8 +7,11 @@
 #ifndef F2C_INCLUDE
 #define F2C_INCLUDE
 
-typedef long int integer;
-typedef unsigned long int uinteger;
+// typedef long int integer;
+// typedef unsigned long int uinteger;
+typedef int integer;
+typedef unsigned int uinteger;
+
 typedef char *address;
 typedef short int shortint;
 typedef float real;
@@ -42,9 +45,12 @@ typedef short flag;
 typedef short ftnlen;
 typedef short ftnint;
 #else
-typedef long int flag;
-typedef long int ftnlen;
-typedef long int ftnint;
+// typedef long int flag;
+// typedef long int ftnlen;
+// typedef long int ftnint;
+typedef int flag;
+typedef int ftnlen;
+typedef int ftnint;
 #endif
 
 /*external read, write*/
